# [oneDNN] Parallelize UnsortedSegmentOp by simpler method