--- a/tensorflow/core/kernels/segment_reduction_ops_test.cc
+++ b/tensorflow/core/kernels/segment_reduction_ops_test.cc
@@ -38,6 +38,74 @@ limitations under the License.
 
 namespace tensorflow {
 
+static void BM_UnsortedSegmentReduction(::testing::benchmark::State& state,
+                                        const string& reduction, int num_rows,
+                                        int num_cols, int segment_size) {
+  std::unique_ptr<Device> device(
+      DeviceFactory::NewDevice("CPU", {}, "/job:a/replica:0/task:0"));
+
+  // Create inputs
+  gtl::InlinedVector<TensorValue, 4> reduction_inputs;
+  TensorShape shape1({num_rows, num_cols});
+  Tensor input(DT_FLOAT, shape1);
+  // input.flat<float>().setRandom();
+  reduction_inputs.push_back({nullptr, &input});
+
+  TensorShape shape2({num_rows});
+  Tensor indices(DT_INT32, shape2);
+  test::FillFn<int>(&indices,
+                    [&segment_size](int i) -> int { return i % segment_size; });
+  reduction_inputs.push_back({nullptr, &indices});
+
+  Tensor num_segments(DT_INT32, TensorShape({}));
+  num_segments.scalar<int>()() = segment_size;
+  reduction_inputs.push_back({nullptr, &num_segments});
+
+  NodeDef reduction_node_def;
+  TF_CHECK_OK(NodeDefBuilder(reduction, reduction)
+                  .Input(FakeInput(DT_FLOAT))
+                  .Input(FakeInput(DT_INT32))
+                  .Input(FakeInput(DT_INT32))
+                  .Finalize(&reduction_node_def));
+  Status status;
+  std::unique_ptr<OpKernel> reduction_op(
+      CreateOpKernel(DEVICE_CPU, device.get(), cpu_allocator(),
+                     reduction_node_def, TF_GRAPH_DEF_VERSION, &status));
+
+  OpKernelContext::Params params;
+  params.device = device.get();
+  params.frame_iter = FrameAndIter(0, 0);
+  params.inputs = &reduction_inputs;
+  params.op_kernel = reduction_op.get();
+  std::vector<AllocatorAttributes> attrs;
+  test::SetOutputAttrs(&params, &attrs);
+
+  std::unique_ptr<OpKernelContext> reduction_context(
+      new OpKernelContext(&params));
+
+  reduction_op->Compute(reduction_context.get());
+  TF_CHECK_OK(reduction_context->status());
+  for (auto s : state) {
+    delete reduction_context->release_output(0).tensor;
+    reduction_op->Compute(reduction_context.get());
+  }
+  int64_t bytes_per_iter =
+      static_cast<int64_t>(num_rows * num_cols * sizeof(float));
+  state.SetBytesProcessed(bytes_per_iter * state.iterations());
+}
+
+#define BM_UnsortedReduce(O, R, C, S)                                        \
+  static void BM_##O##_##R##_##C##_##S(::testing::benchmark::State& state) { \
+    BM_UnsortedSegmentReduction(state, #O, R, C, S);                         \
+  }                                                                          \
+  BENCHMARK(BM_##O##_##R##_##C##_##S);
+
+#define BM_UnsortedReduce_Arg(R, C, S) \
+  BM_UnsortedReduce(UnsortedSegmentSum, R, C, S);
+
+BM_UnsortedReduce_Arg(4096, 1024, 1);
+BM_UnsortedReduce_Arg(4096, 1024, 128);
+
 template <typename Index>
 static void BM_SegmentReduction(::testing::benchmark::State& state,
                                 const string& reduction, Index num_rows,