This PR is a simpler version of the reverted PR: https://github.com/tensorflow/tensorflow/pull/49152

The previous PR **used a complex method to balance workload** for better parallelism efficiency, but it's failed in internal test and reverted automatically. I removed the balance policy to make it simpler and safer. The new implementation can also get good performance if the workload is balance naturally. Even in the imbalance scenario it won't be worse than original single thread implementation.

The new method is just to parallelize `UnsortedSegmentOp` by each output row:
```
//   input   segment_ids                 num_segments  operation
//   | a0 |  | 0 |            worker 1:  |0|           f(a0, a1)
//   | b0 |  | 1 |            worker 2:  |1|           f(b0, b1)
// N | c0 |  | 2 |       -->  worker 3:  |2|           f(c0)
//   | b1 |  | 1 |
//   | a1 |  | 0 |
```

Single op improvement in TF benchmark: **1.92x ~ 14.46x**
Before opt:
```
--------------------------------------------------------------------------------------------------------------
Benchmark                                                    Time             CPU   Iterations UserCounters...
--------------------------------------------------------------------------------------------------------------
BM_UnsortedSegmentSum_4096_1024_1                       520890 ns       520870 ns         1344 bytes_per_second=29.9979G/s
BM_UnsortedSegmentSum_4096_1024_128                     700895 ns       700889 ns         1019 bytes_per_second=22.2931G/s
```
After opt:
```
--------------------------------------------------------------------------------------------------------------
Benchmark                                                    Time             CPU   Iterations UserCounters...
--------------------------------------------------------------------------------------------------------------
BM_UnsortedSegmentSum_4096_1024_1                       270874 ns       270875 ns         2586 bytes_per_second=57.6835G/s
BM_UnsortedSegmentSum_4096_1024_128                     115952 ns        48465 ns        14288 bytes_per_second=322.396G/s
```

We also see more than 5% throughput improvement when run public recommendation model on CPU with this OPT.

Signed-off-by: Lu Teng [teng.lu@intel.com](mailto:teng.lu@intel.com)