@Zantares 2022-02-14
Hi @ezhulenev , can you help to review this simpler version since you have checked the original PR?